+++ pdtable/test/io/test__excel.py
@@ -1,5 +1,6 @@
 import datetime
 
+import pytest
 import pandas as pd
 import openpyxl
 
@@ -11,7 +12,14 @@ except ImportError:
 
 from pdtable import Table
 from pdtable.io.excel import write_excel
-from pdtable.io._excel_openpyxl import _append_table_to_openpyxl_worksheet
+from pdtable.io._excel_openpyxl import _append_table_to_openpyxl_worksheet, deep_get
+
+
+def test_deep_get():
+    d = {'meta': {'status': 'OK', 'status_code': 200}}
+    assert deep_get(d, ['meta', 'status_code']) == 200
+    assert deep_get(d, ['garbage', 'status_code']) is None
+    assert deep_get(d, ['meta', 'garbage'], default='-') == '-'
 
 
 def test__append_table_to_openpyxl_worksheet():
@@ -30,7 +38,7 @@ def test__append_table_to_openpyxl_worksheet():
     ws = wb.active
 
     # Act
-    _append_table_to_openpyxl_worksheet(t, ws, num_blank_rows_between_tables=1)
+    _append_table_to_openpyxl_worksheet(t, ws, sep_lines=1)
 
     # Assert worksheet looks as expected:
     # table header by row
@@ -142,7 +150,7 @@ def test_write_excel__multiple_sheets(tmp_path):
     out_path.unlink()
 
 
-def test_write_excel_with_formatting(tmp_path):
+def test_write_excel__style(tmp_path):
     # Make a couple of tables
     t = Table(name="foo")
     t["place"] = ["home", "work", "beach", "wonderland"]
@@ -168,7 +176,7 @@ def test_write_excel_with_formatting(tmp_path):
 
     # Write tables to workbook, save, and re-load
     out_path = tmp_path / "foo.xlsx"
-    write_excel([t, t2, t3], out_path, style=True)
+    write_excel([t, t2, t3], out_path, styles=True)
     wb = openpyxl.load_workbook(out_path)
     ws = wb.active
 
@@ -269,7 +277,154 @@ def test_write_excel_with_formatting(tmp_path):
     out_path.unlink()
 
 
-def test_write_excel_with_formatting_and_2_blank_rows_between_tables(tmp_path):
+def test_write_excel__custom_style(tmp_path):
+    # Make a table
+    t = Table(name="foo")
+    t["place"] = ["home", "work", "beach", "wonderland"]
+    t.add_column("distance", list(range(3)) + [float("nan")], "km")
+    nc = len(t.column_names)
+    nr = len(t.df)
+
+    # Make a style specification as a JSON-like data structure
+    style_spec = {
+        "table_name": {
+            "font": {
+                "name": "Times New Roman",
+                "size": 24,
+                "color": "FF0000",   # RGB hex color code
+                "bold": True,
+            },
+            "fill": {
+                "color": "00AAAAAA",  # leading 'aa' transparency values are accepted (but unused)
+            },
+        },
+        "destinations": {
+            "font": {
+                "italic": True,
+                "color": "0000FF",
+            },
+            "fill": {
+                "color": "888888",
+            },
+        },
+        "column_names": {
+            "font": {
+                "color": "444400",
+                "bold": True,
+            },
+            "fill": {
+                "color": "777777",
+            },
+        },
+        "units": {
+            "font": {
+                "color": "440044",
+            },  # --------------------- fill unspecified, leave untouched
+        },
+        "values": {
+            "alignment": {
+                "horizontal": "left",
+            },
+            "fill": {
+                "color": "EEEEEE",
+            },  # --------------------- font unspecified, leave untouched
+        },
+    }
+
+    # Write tables to workbook, save, and re-load
+    out_path = tmp_path / "foo_custom_style.xlsx"
+    write_excel([t], out_path, styles=style_spec)
+    wb = openpyxl.load_workbook(out_path)
+    ws = wb.active
+
+    # Check table formatting
+    # table name
+    assert ws["A1"].fill.fill_type == "solid"
+    assert ws["A1"].fill.start_color.value == "00AAAAAA"
+    assert ws["A1"].font.size == 24
+    assert ws["A1"].font.color.value == "00FF0000"
+    assert ws["A1"].font.name == "Times New Roman"
+    assert ws["A1"].font.bold is True
+
+    # destinations
+    assert ws["A2"].fill.fill_type == "solid"
+    assert ws["A2"].fill.start_color.value == "00888888"
+    assert ws["A2"].font.color.value == "000000FF"
+    assert ws["A2"].font.bold is False
+    assert ws["A2"].font.italic is True
+
+    # column names
+    assert [ws.cell(3, c).fill.fill_type for c in range(1, nc+1)] == ["solid"] * nc
+    assert [ws.cell(3, c).fill.start_color.value for c in range(1, nc+1)] == ["00777777"] * nc
+    assert [ws.cell(3, c).font.color.value for c in range(1, nc+1)] == ["00444400"] * nc
+    assert [ws.cell(3, c).font.bold for c in range(1, nc+1)] == [True] * nc
+
+    # column units
+    assert [ws.cell(4, c).fill.fill_type for c in range(1, nc+1)] == [None] * nc  # left as default
+    assert [ws.cell(4, c).font.color.value for c in range(1, nc+1)] == ["00440044"] * nc
+    assert [ws.cell(4, c).font.bold for c in range(1, nc+1)] == [False] * nc
+    assert [ws.cell(4, c).alignment.horizontal for c in range(1, nc+1)] == [None] * nc
+
+    # values
+    assert [[ws.cell(4 + r, c).fill.fill_type for c in range(1, nc + 1)] for r in range(1, nr + 1)] == [["solid"] * nc] * nr
+    assert [[ws.cell(4 + r, c).fill.start_color.value for c in range(1, nc + 1)] for r in range(1, nr + 1)] == [["00EEEEEE"] * nc] * nr
+    assert [[ws.cell(4 + r, c).alignment.horizontal for c in range(1, nc + 1)] for r in range(1, nr + 1)] == [["left"] * nc] * nr
+
+
+@pytest.mark.parametrize(
+    "err_msg_match,style_spec",
+    [
+        ("Invalid.*table_name", {"table_name": {"font": {"size": "NOT_A_NUMBER!!!"}}}),
+        ("Invalid.*table_name", {"table_name": {"font": {"color": "NOT_A_COLOR!!!"}}}),
+        ("Invalid.*destinations", {"destinations": {"fill": {"color": "NOT_A_COLOR!!!"}}}),
+        ("Invalid.*values", {"values": {"alignment": {"horizontal": "FOOBAR!!!"}}}),
+    ]
+)
+def test_write_excel__raises_error_on_invalid_style_spec(tmp_path, err_msg_match, style_spec):
+    # Make a table
+    t = Table(name="foo")
+    t["place"] = ["home", "work", "beach", "wonderland"]
+    t.add_column("distance", list(range(3)) + [float("nan")], "km")
+
+    # Invalid font size
+    with pytest.raises(ValueError, match=err_msg_match):
+        write_excel([t], tmp_path / "foo_invalid_style.xlsx", styles=style_spec)
+
+
+def test_write_excel__transposed_table_units_and_values_are_centered_by_default(tmp_path):
+    # Make a table
+    t = Table(name="foo")
+    t["place"] = ["home", "work", "beach", "wonderland"]
+    t.add_column("distance", list(range(3)) + [float("nan")], "km")
+    t.metadata.transposed = True
+    nc = len(t.column_names)
+    nr = len(t.df)
+
+    # DEFAULT ALIGNMENT: CENTER
+    # Write tables to workbook with default styles, save, and re-load
+    out_path = tmp_path / "foo_custom_style.xlsx"
+    write_excel([t], out_path, styles=True)  # <<< Default style
+    wb = openpyxl.load_workbook(out_path)
+    ws = wb.active
+
+    # column units and values are centered
+    assert [ws.cell(2 + c, 2).alignment.horizontal for c in range(1, nc+1)] == ["center"] * nc
+    assert [[ws.cell(2 + c, 2 + r).alignment.horizontal for c in range(1, nc + 1)] for r in range(1, nr + 1)] == [["center"] * nc] * nr
+
+    # DEFAULT ALIGNMENT NOT APPLIED when custom alignment is specified
+    # Write tables to workbook with custom alignment styles, save, and re-load
+    out_path = tmp_path / "foo_custom_style.xlsx"
+    left = {"alignment": {"horizontal": "left"}}
+    write_excel([t], out_path, styles={"units": left, "values": left})  # << Custom alignment
+    wb = openpyxl.load_workbook(out_path)
+    ws = wb.active
+
+    # column units and values are not centered
+    assert [ws.cell(2 + c, 2).alignment.horizontal for c in range(1, nc+1)] == ["left"] * nc
+    assert [[ws.cell(2 + c, 2 + r).alignment.horizontal for c in range(1, nc + 1)] for r in range(1, nr + 1)] == [["left"] * nc] * nr
+
+
+def test_write_excel__sep_lines(tmp_path):
     # Make a couple of tables
     t = Table(name="foo")
     t["place"] = ["home", "work", "beach", "wonderland"]
@@ -295,11 +450,11 @@ def test_write_excel_with_formatting_and_2_blank_rows_between_tables(tmp_path):
 
     # Write tables to workbook, save, and re-load
     out_path = tmp_path / "foo.xlsx"
-    write_excel([t, t2, t3], out_path, style=True, num_blank_rows_between_tables=2)
+    write_excel([t, t2, t3], out_path, sep_lines=2)
     wb = openpyxl.load_workbook(out_path)
     ws = wb.active
 
-    # Tables are started as expected
+    # Tables start on the expected rows
     assert ws["A1"].value == "**foo"
     assert ws["A11"].value == "**bar*"
     assert ws["A17"].value == "**bas*"
@@ -311,7 +466,7 @@ def test_read_write_excel__round_trip_with_styles(tmp_path):
     bundle = TableBundle(read_excel("pdtable/test/io/input/foo.xlsx"))
     out_path = tmp_path / "foo_styled.xlsx"
     # Doesn't crash on write
-    write_excel(bundle, out_path, style=True)
+    write_excel(bundle, out_path, styles=True)
     # Re-read bundle is same as first one
     bundle2 = TableBundle(read_excel(out_path))
     for t, t2 in zip(bundle, bundle2):